minor fixes
git-svn-id: http://svn.xp-dev.com/svn/wxextension@3377 f22100f3-aa73-48fe-a4fb-fd497bb32605

+++ trunk/extension/src/report/listview.cpp
@@ -883,7 +883,7 @@ void wxExListViewWithFrame::RunItems(const wxExTool& tool)
     return;
   }
 
-  int i = -1;
+  long i = -1;
 
   wxExStatistics<long> stats;
